fix it some problem & setting time & with sqlite3

+++ db/migrate/20181017124252_create_weights.rb
@@ -3,9 +3,9 @@ class CreateWeights < ActiveRecord::Migration[5.2]
     create_table :weights do |t|
       t.date :date
       t.string :gender
-      t.decimal :lenght
-      t.decimal :weight
-      t.decimal :ideal_weight
+      t.float :lenght
+      t.float :weight
+      t.float :ideal_weight
 
       t.timestamps
     end
